# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MP3 Downloader API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * The source platform
 */
export type DownloadRequestSource = typeof DownloadRequestSource[keyof typeof DownloadRequestSource];


export const DownloadRequestSource = {
  spotify: 'spotify',
  youtube: 'youtube',
} as const;

export interface DownloadRequest {
  /** The Spotify or YouTube URL to download */
  url: string;
  /** The source platform */
  source: DownloadRequestSource;
}

export type DownloadJobStatus = typeof DownloadJobStatus[keyof typeof DownloadJobStatus];


export const DownloadJobStatus = {
  pending: 'pending',
  processing: 'processing',
  done: 'done',
  error: 'error',
} as const;

export type DownloadJobSource = typeof DownloadJobSource[keyof typeof DownloadJobSource];


export const DownloadJobSource = {
  spotify: 'spotify',
  youtube: 'youtube',
} as const;

export interface DownloadJob {
  jobId: string;
  status: DownloadJobStatus;
  source: DownloadJobSource;
  url: string;
  /** @nullable */
  filename?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  createdAt?: string;
}

export interface ErrorResponse {
  error: string;
}