# lib/api-zod/src/generated/types/downloadJob.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MP3 Downloader API
 * OpenAPI spec version: 0.1.0
 */
import type { DownloadJobSource } from './downloadJobSource';
import type { DownloadJobStatus } from './downloadJobStatus';

export interface DownloadJob {
  jobId: string;
  status: DownloadJobStatus;
  source: DownloadJobSource;
  url: string;
  /** @nullable */
  filename?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  createdAt?: string;
}