# lib/api-zod/src/generated/types/downloadRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MP3 Downloader API
 * OpenAPI spec version: 0.1.0
 */
import type { DownloadRequestSource } from './downloadRequestSource';

export interface DownloadRequest {
  /** The Spotify or YouTube URL to download */
  url: string;
  /** The source platform */
  source: DownloadRequestSource;
}